Relocate the class extensions
